Move current code to internal/pppoe.
Also create a testutil internal package to handle all the docker
nonsense for future packages.
